Moved RecordStore to a separate directory
