+++ slatedb/src/lib.rs
@@ -102,6 +102,7 @@ mod paths;
 mod proptest_util;
 mod rand;
 mod reader;
+mod record;
 mod retention_iterator;
 mod retrying_object_store;
 mod row_codec;
